+++ common/routes/authtoken.js
@@ -3,20 +3,19 @@
  * Created by kc on 29.01.15.
  */
 
-const express = require('express');
-const router  = express.Router();
-const session = require('express-session');
-const uuid    = require('uuid/v4');
-const logger  = require('../lib/logger').getLogger('authToken');
+const express    = require('express');
+const router     = express.Router();
+const session    = require('express-session');
+const logger     = require('../lib/logger').getLogger('authToken');
+const {v4: uuid} = require('uuid');
 
 /* GET the authtoken, which you only can get when logged in */
 router.get('/', function (req, res) {
 
-  
+
   if (req.session.authToken) {
     logger.info(`Auth token REFRESH for ${req.session.passport.user}: ${req.session.authToken}`);
-  }
-  else {
+  } else {
     req.session.authToken = uuid();
     logger.info(`NEW auth token for ${req.session.passport.user}: ${req.session.authToken}`);
   }